+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridge.kt
@@ -138,6 +138,7 @@ open class MiniAppMessageBridge {
             ActionType.GET_USER_NAME.action -> userInfoBridge.onGetUserName(callbackObj.id)
             ActionType.GET_PROFILE_PHOTO.action -> userInfoBridge.onGetProfilePhoto(callbackObj.id)
             ActionType.GET_ACCESS_TOKEN.action -> userInfoBridge.onGetAccessToken(callbackObj)
+            ActionType.GET_POINTS.action -> userInfoBridge.onGetPoints(callbackObj.id)
             ActionType.SET_SCREEN_ORIENTATION.action -> screenBridgeDispatcher.onScreenRequest(callbackObj)
             ActionType.GET_CONTACTS.action -> userInfoBridge.onGetContacts(callbackObj.id)
             ActionType.SEND_MESSAGE_TO_CONTACT.action -> chatBridge.onSendMessageToContact(
@@ -173,7 +174,11 @@ open class MiniAppMessageBridge {
         val successCallback = { uniqueId: String ->
             bridgeExecutor.postValue(callbackObj.id, uniqueId)
         }
-        val errorCallback = { message: String -> throw MiniAppSdkException(message) }
+        val errorCallback = { message: String ->
+            bridgeExecutor.postError(
+                    callbackObj.id, "${ErrorBridgeMessage.ERR_UNIQUE_ID} $message"
+            )
+        }
 
         getUniqueId(successCallback, errorCallback)
     } catch (e: Exception) {