+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/SettingsMenuActivity.kt
@@ -24,9 +24,7 @@ import com.rakuten.tech.mobile.testapp.ui.input.MiniAppInputActivity
 import com.rakuten.tech.mobile.testapp.ui.miniapplist.MiniAppListActivity
 import com.rakuten.tech.mobile.testapp.ui.permission.MiniAppDownloadedListActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.MenuBaseActivity.Companion.MENU_SCREEN_NAME
-import com.rakuten.tech.mobile.testapp.ui.userdata.AccessTokenActivity
-import com.rakuten.tech.mobile.testapp.ui.userdata.ContactListActivity
-import com.rakuten.tech.mobile.testapp.ui.userdata.ProfileSettingsActivity
+import com.rakuten.tech.mobile.testapp.ui.userdata.*
 import kotlinx.coroutines.launch
 import java.net.URL
 import kotlin.properties.Delegates
@@ -122,7 +120,17 @@ class SettingsMenuActivity : BaseActivity() {
             MiniAppDownloadedListActivity.start(this@SettingsMenuActivity)
         }
 
-        binding.buttonAccessToken.setOnClickListener { AccessTokenActivity.start(this@SettingsMenuActivity) }
+        binding.buttonAccessToken.setOnClickListener {
+            AccessTokenActivity.start(this@SettingsMenuActivity)
+        }
+
+        binding.buttonPoints.setOnClickListener {
+            PointsActivity.start(this@SettingsMenuActivity)
+        }
+
+        binding.buttonQA.setOnClickListener {
+            QASettingsActivity.start(this@SettingsMenuActivity)
+        }
 
         validateInputIDs()
     }